requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-debug-toolbar==4.4.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
factory-boy==3.3.0
Faker==25.3.0
idna==3.7
Markdown==3.6
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
urllib3==2.2.1
